Replay plugin: move the legacy implementation to a different file
So that the new implementation can live in a more permanent and
obvious location.
